backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1

# Geospatial
shapely==2.0.2
pyproj==3.6.1

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Message Queue
pika==1.3.2

# AWS S3
boto3==1.29.7
botocore==1.32.7

# External APIs
twilio==8.10.3
requests==2.31.0
httpx==0.25.2
sendgrid==6.11.0

# OpenAI for NLP
openai==1.3.7

# Weather APIs
pyowm==3.3.0

# ML/AI
numpy==1.26.2
pillow==10.1.0
scikit-learn==1.3.2

# WebSocket
python-socketio==5.10.0
websockets==12.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
phonenumbers==8.13.26

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
slowapi==0.1.9

# Supabase (Optional - for enhanced features)
supabase>=2.0.0
